# Small changes for flexibility and performance