@csteinmetz1 2024-02-09
Thanks @simonschwaer ! I would love to merge these changes. I think everything is good that doesn't cause any change in default behavior. So, would it be possible to split change 2. dealing with `eps` to a separate PR?